+++ CHANGELOG.md
@@ -37,6 +37,8 @@ however, insignificant breaking changes does not guarantee a major version bump,
 - Added `?plugins update` for updating all installed plugins.
 - Reintroduce flake8 and use bandit for security issues detection.
 - Add travis checks for 3.6 in Linux and 3.7 for MacOS and Windows.
+- Eval commands are logged in debug logs.
+- Presence updates 30 minutes instead of 45 now.
 
 ### Internal
 
@@ -44,6 +46,7 @@ however, insignificant breaking changes does not guarantee a major version bump,
   - To get/set the raw value, access through `config[]`.
 - Prerelease naming scheme is now `x.x.x-devN`.
 - `trigger_typing` has been moved to `core.utils.trigger_typing`, original location is deprecated.
+- Simpler status and activity logic.
 
 # v3.2.2